+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - (Feature) EnforcedResignLeadership action
 - (Maintenance) Make scale_down_candidate annotation obsolete
 - (Bugfix) Fix ResignJob ID propagation
+- (Bugfix) Allow shards with RF1 in EnforcedResignLeadership action
  
 ## [1.2.33](https://github.com/arangodb/kube-arangodb/tree/1.2.33) (2023-09-27)
 - (Maintenance) Bump golang.org/x/net to v0.13.0